+++ website/src/components/campaign/campaigns-overview.tsx
@@ -73,6 +73,8 @@ export const CampaignsOverview = async ({
 		programPlaceholder: translator.t('campaigns-page.submission.program-placeholder'),
 		currencyPlaceholder: translator.t('campaigns-page.submission.currency-placeholder'),
 		imageHint: translator.t('campaigns-page.submission.image-hint'),
+		continue: translator.t('campaigns-page.submission.continue'),
+		back: translator.t('campaigns-page.submission.back'),
 		errors: submissionErrors,
 	};
