+++ website/src/lib/i18n/locales/de/website-common.json
@@ -199,6 +199,8 @@
 			"program-placeholder": "Programm auswählen",
 			"currency-placeholder": "Währung auswählen",
 			"image-hint": "JPEG, PNG oder WebP bis 5 MB.",
+			"continue": "Weiter",
+			"back": "Zurück",
 			"errors": {
 				"title-required": "Titel ist erforderlich.",
 				"title-too-long": "Titel ist zu lang.",
