+++ website/src/lib/i18n/locales/en/website-common.json
@@ -199,6 +199,8 @@
 			"program-placeholder": "Select a program",
 			"currency-placeholder": "Select a currency",
 			"image-hint": "JPEG, PNG, or WebP up to 5 MB.",
+			"continue": "Continue",
+			"back": "Back",
 			"errors": {
 				"title-required": "Title is required.",
 				"title-too-long": "Title is too long.",
